Clean up and reorganization
