+++ src/Services/DataProcessing/API/Controllers/DataProcessingController.cs
@@ -1,4 +1,4 @@
-using DataProcessing.Application.Dtos;
+using DataProcessing.Application.Dtos.LogErrorTimeData;
 using DataProcessing.Application.Queries.Interfaces;
 using DataProcessing.Application.Commands.Interfaces;
 using Microsoft.AspNetCore.Mvc;
