+++ src/Services/DataProcessing/Application/Dtos/LogErrorTimeData/LogProcessRequestDto.cs
@@ -1,4 +1,4 @@
-namespace DataProcessing.Application.Dtos;
+namespace DataProcessing.Application.Dtos.LogErrorTimeData;
 public class LogProcessRequestDto
 {
     public string Content { get; private set; }
