+++ src/Services/DataProcessing/Application/Dtos/LogErrorTimeData/LogSendToAnalysisDto.cs
@@ -1,4 +1,4 @@
-namespace DataProcessing.Application.Dtos;
+namespace DataProcessing.Application.Dtos.LogErrorTimeData;
 public class LogSendToAnalysisDto
 {
     public string MLType { get; set; } // The type of dataset the log is processed into to be used for ML
