+++ src/Services/ML/API/app.py
@@ -8,7 +8,7 @@ def dataset_collector():
         # Get the string CSV data from the request
         csv_data = request.data.decode('utf-8')
         # Save the CSV data to a file
-        with open('../Data/Datasets/dataset.csv', 'w') as f:
+        with open('../Models/LogErrorTime/Data/Datasets/dataset.csv', 'w') as f:
             f.write(csv_data)
         print(csv_data)
         return 'Dataset saved successfully'
@@ -19,7 +19,7 @@ def dataset_collector():
 def dataset():
     if request.method == 'GET':
         # Read the CSV data from the file
-        with open('../Data/Datasets/dataset.csv', 'r') as f:
+        with open('../Models/LogErrorTime/Data/Datasets/dataset.csv', 'r') as f:
             csv_data = f.read()
         return csv_data
